pyproject.toml
[project]
name = "studysync"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.74.1",
    "fastapi>=0.121.3",
    "google-adk>=1.18.0",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.4",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
